refactor: create GXCloudStoreBackend
